Move names out of the helper package 🌿
