+++ pkg/helper/deleter/deleter.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 
 	"github.com/tektoncd/cli/pkg/cli"
-	"github.com/tektoncd/cli/pkg/helper/names"
+	"github.com/tektoncd/cli/pkg/names"
 	"go.uber.org/multierr"
 )
